pyproject.toml
[project]
name = "devagent"
version = "0.1.0"
description = "AI-powered tutorial generator using multi-agent workflows to create Getting Started guides for any library."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "chromadb>=1.4.1",
    "httpx>=0.27.0",
    "langchain>=0.3.0",
    "langchain-openai>=1.1.7",
    "langfuse>=3.0.0",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.6",
    "llama-index>=0.14.12",
    "llama-index-embeddings-google-genai>=0.3.2",
    "llama-index-readers-web>=0.5.6",
    "llama-index-vector-stores-chroma>=0.5.5",
    "mcp>=1.26.0",
    "packaging>=24.0",
    "pydantic>=2.0.0",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.0.0",
    "streamlit>=1.40.0",
    "tavily-python>=0.7.19",
    "trafilatura>=1.12.0",
    "tree-sitter>=0.21.3",
    "tree-sitter-languages>=1.10.2",
    "tree-sitter-language-pack>=0.13.0",
    "huggingface-hub>=1.4.0",
    "huggingface>=0.0.1",
    "llama-index-embeddings-huggingface-api>=0.2.0",
]
